+++ CMakeLists.txt
@@ -1,3 +1,5 @@
+cmake_minimum_required(VERSION 3.18)
+
 project(TroyNew LANGUAGES C CXX CUDA)
 
 set(CMAKE_CXX_STANDARD 17)
